Added rejoin option to merge command
If the commits splitted from a subtree needs some interactive rebasing or other alteration beore being used in another place then an immediate rejoining isn't possible.

Currently there's no way of doing something similar to a rejoin if you don't do it as part of a split. This commit adds a rejoin option to the merge command so that a rejoin merge can be made after alterations have been made to the splitted out commits.